Rename CompactionStrategy to RewriteStrategy
Remove all remaining chunk references
